Initial GreyVest theme refactor
