+++ src/themes/greyVest/HighlightWrap.js
@@ -1,0 +1,5 @@
+import React from 'react'
+
+let HighlightWrap = ({ props, style }) => <b style={{ backgroundColor: 'yellow', ...style }} {...props} />
+
+export default HighlightWrap
