+++ src/themes/greyVest/Modal.js
@@ -1,0 +1,6 @@
+import DefaultModal from '../../layout/Modal'
+import { defaultProps } from 'recompose'
+
+let Modal = defaultProps({ className: 'gv-body' })(DefaultModal)
+
+export default Modal
