+++ src/themes/greyVest/PagerItem.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import { observer } from 'mobx-react'
+
+let PagerItem = ({ active, disabled, ...props }) => (
+  <span
+    className={`gv-pager-item ${disabled ? 'disabled' : ''} ${
+      active ? 'active' : ''
+    }`}
+    {...props}
+  />
+)
+
+export default observer(PagerItem)
